Refactor to have generic i2c connection class
Refactored the code to provide a generic i2c connection to a single
device. This way we can share the functionality between other i2c
devices down the road. I could see an argument for defining a pure
virtual interface and then a linux implementation from that, but that
can wait.
