+++ mpl3115a2.cpp
@@ -9,13 +9,6 @@
 #include <thread>
 #include <vector>
 
-#include <errno.h>
-#include <fcntl.h>
-#include <linux/i2c.h>
-#include <linux/i2c-dev.h>
-#include <sys/ioctl.h>
-#include <unistd.h>
-
 #include "mpl3115a2.hpp"
 
 
@@ -46,42 +39,11 @@ constexpr uint8_t STATUS_PDR_MASK =  0x04;
 constexpr uint8_t STATUS_PTDR_MASK = 0x08;
 
 
-MPL3115A2::MPL3115A2(const unsigned int adapterNumber)
+MPL3115A2::MPL3115A2(const unsigned int adapterNumber) :
+    m_connection(new I2cAbstraction(adapterNumber, MPL3115A2_ADDRESS))
 {
-    std::ostringstream oss;
-    oss << "/dev/i2c-" << adapterNumber;
-    m_i2cFilename = oss.str();
-    m_i2cFile = open(m_i2cFilename.c_str(), O_RDWR);
-    if (m_i2cFile < 0)
-    {
-        std::ostringstream err;
-        err << "Could not open " << m_i2cFilename << std::endl << strerror(errno);
-        throw std::runtime_error(err.str());
-    }
-    if (ioctl(m_i2cFile, I2C_SLAVE, MPL3115A2_ADDRESS) < 0)
-    {
-        std::ostringstream err;
-        err << "Could not set slave address\n" << strerror(errno);
-        throw std::runtime_error(err.str());
-    }
-    unsigned long funcs = 0;
-    if (ioctl(m_i2cFile, I2C_FUNCS, &funcs) < 0)
-    {
-        std::ostringstream err;
-        err << "Could not get available functionality from the i2c adapter"
-            << std::endl << strerror(errno);
-        throw std::runtime_error(err.str());
-    }
-    if (!(funcs & I2C_FUNC_I2C))
-    {
-        std::ostringstream err;
-        err << "The adapter does not support a mixed read write transaction"
-            << std::endl << strerror(errno);
-        throw std::runtime_error(err.str());
-    }
-
     // Confirm that the device at this address is indeed the MPL3115A2
-    uint8_t whoIsThis = readBytes(WHO_AM_I, 1)[0];
+    uint8_t whoIsThis = m_connection->readBytes(WHO_AM_I, 1)[0];
     if (whoIsThis != DEVICE_ID)
     {
         std::ostringstream err;
@@ -92,8 +54,8 @@ MPL3115A2::MPL3115A2(const unsigned int adapterNumber)
     }
     else
     {
-        std::cout << "MPL3115A2 confirmed to be on I2C bus represented by "
-                  << m_i2cFilename << std::endl;
+        std::cout << "MPL3115A2 confirmed to be on I2C bus adapter "
+                  << adapterNumber << std::endl;
     }
 
     configureDataReadyFlag();
@@ -105,8 +67,8 @@ uint8_t MPL3115A2::enterStandbyMode(void) const
 {
     // Enter standby mode to allow register writing
     // Return the register data to allow it to be restored to previous state
-    uint8_t controlRegisterData = readBytes(CTRL_REG1, 1)[0];
-    writeByte(CTRL_REG1, controlRegisterData & ~STANDBY_BAR_MASK);  // Clear standby bar bit
+    uint8_t controlRegisterData = m_connection->readBytes(CTRL_REG1, 1)[0];
+    m_connection->writeByte(CTRL_REG1, controlRegisterData & ~STANDBY_BAR_MASK);  // Clear standby bar bit
     return controlRegisterData;
 }
 
@@ -117,7 +79,7 @@ void MPL3115A2::configureAltimeterMode(void)
     uint8_t controlRegisterData = enterStandbyMode();
 
     // Set the mode to altimeter and set the standby-bar bit to deactivate standby mode
-    writeByte(CTRL_REG1, controlRegisterData | ALTIMETER_MASK | STANDBY_BAR_MASK);
+    m_connection->writeByte(CTRL_REG1, controlRegisterData | ALTIMETER_MASK | STANDBY_BAR_MASK);
     isAltimeterMode = true;
     isBarometerMode = false;
 }
@@ -129,7 +91,7 @@ void MPL3115A2::configureBarometerMode(void)
     uint8_t controlRegisterData = enterStandbyMode();
 
     // Set the mode to barometer and set the standby-bar bit to deactivate standby mode
-    writeByte(CTRL_REG1, (controlRegisterData & ~ALTIMETER_MASK) | STANDBY_BAR_MASK);
+    m_connection->writeByte(CTRL_REG1, (controlRegisterData & ~ALTIMETER_MASK) | STANDBY_BAR_MASK);
     isAltimeterMode = false;
     isBarometerMode = true;
 }
@@ -142,67 +104,10 @@ void MPL3115A2::configureDataReadyFlag(void) const
     uint8_t controlRegisterData = enterStandbyMode();
 
     // Configure the sensor data register to raise a status flag when any new data is available
-    writeByte(PT_DATA_CFG, PT_DATA_CFG_DREM_MASK | PT_DATA_CFG_TDEFE_MASK | PT_DATA_CFG_PDEFE_MASK);
+    m_connection->writeByte(PT_DATA_CFG, PT_DATA_CFG_DREM_MASK | PT_DATA_CFG_TDEFE_MASK | PT_DATA_CFG_PDEFE_MASK);
 
     // Set device back to active mode
-    writeByte(CTRL_REG1, controlRegisterData | STANDBY_BAR_MASK);
-}
-
-
-std::vector<uint8_t> MPL3115A2::readBytes(uint8_t reg, unsigned int size) const
-{
-    struct i2c_rdwr_ioctl_data packagedMessages;
-    struct i2c_msg messages[2];
-
-    // This message is responsible for telling the sensor which register we want data from
-    messages[0].addr = MPL3115A2_ADDRESS;
-    messages[0].flags = 0;
-    messages[0].len = sizeof(reg);
-    messages[0].buf = &reg;
-
-    // This message contains the data from the register
-    std::unique_ptr<uint8_t[]> data(new uint8_t[size]);
-    messages[1].addr  = MPL3115A2_ADDRESS;
-    messages[1].flags = I2C_M_RD;
-    messages[1].len   = size;
-    messages[1].buf   = data.get();
-
-    // Ask for the transaction to take place
-    packagedMessages.msgs = messages;
-    packagedMessages.nmsgs = 2;
-    if(ioctl(m_i2cFile, I2C_RDWR, &packagedMessages) < 0)
-    {
-        std::ostringstream err;
-        err << "Could not perform read" << std::endl << strerror(errno);
-        throw std::runtime_error(err.str());
-    }
-    return std::vector<uint8_t>(data.get(), data.get() + size);
-}
-
-
-void MPL3115A2::writeByte(uint8_t reg, uint8_t data) const
-{
-    struct i2c_rdwr_ioctl_data packagedMessages;
-    struct i2c_msg message;
-
-    // This message contains the register to write to as well as the data to write
-    uint8_t out[2];
-    out[0] = reg;
-    out[1] = data;
-    message.addr = MPL3115A2_ADDRESS;
-    message.flags = 0;
-    message.len = sizeof(out);
-    message.buf = out;
-
-    packagedMessages.msgs = &message;
-    packagedMessages.nmsgs = 1;
-
-    if(ioctl(m_i2cFile, I2C_RDWR, &packagedMessages) < 0)
-    {
-        std::ostringstream err;
-        err << "Could not perform write" << std::endl << strerror(errno);
-        throw std::runtime_error(err.str());
-    }
+    m_connection->writeByte(CTRL_REG1, controlRegisterData | STANDBY_BAR_MASK);
 }
 
 
@@ -214,10 +119,10 @@ MPL3115A2DATA MPL3115A2::getPressure(void)
     }
 
     // Poll until there is data available
-    uint8_t status = readBytes(STATUS, 1)[0];
+    uint8_t status = m_connection->readBytes(STATUS, 1)[0];
     while (!(status & STATUS_PTDR_MASK))
     {
-        status = readBytes(STATUS, 1)[0];
+        status = m_connection->readBytes(STATUS, 1)[0];
         std::chrono::milliseconds timespan(10);
         std::this_thread::sleep_for(timespan);
     }
@@ -258,10 +163,10 @@ MPL3115A2DATA MPL3115A2::getAltitude(void)
     {
         configureAltimeterMode();
     }
-    uint8_t status = readBytes(STATUS, 1)[0];
+    uint8_t status = m_connection->readBytes(STATUS, 1)[0];
     while (!(status & STATUS_PTDR_MASK))
     {
-        status = readBytes(STATUS, 1)[0];
+        status = m_connection->readBytes(STATUS, 1)[0];
         std::chrono::milliseconds timespan(10);
         std::this_thread::sleep_for(timespan);
     }
@@ -318,5 +223,5 @@ double MPL3115A2::calculateTemperature(uint8_t MSB, uint8_t LSB)
 std::vector<uint8_t> MPL3115A2::getData(void) const
 {
     // Get all the data in one transaction
-    return readBytes(PRESSURE_MSB, 5);
+    return m_connection->readBytes(PRESSURE_MSB, 5);
 }
